pyproject.toml
[tool.poetry]
name = "EconomyCalendarSnapshots"
version = "0.1.0"
description = ""
authors = ["maxisoft <maxisoft.eu+git@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cloudscraper = ">0"
orjson = ">0"
GitPython = ">0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
